Animation made by pixel shaders is supported now.
1. Animation made by pixel shaders is supported now.
2. Logical engine is partly available now.
3. Make the codes a bit more standard (refer to the Google style guide).

+++ inc/V6LogicEngine.h
@@ -26,6 +26,6 @@ namespace VoltaEngine{
 		// 其实应该作封装。
 		static LogicManager Manager;
 		// 全局时间。
-		static int g_time;
+		static long long g_time;
 	};
 }
